requirements.txt
rdflib>=4.2.2
pytest>=4.2.1
networkx>=2.4
greenery
scipy>=1.1.0
antlr4-python3-runtime